Adding a fill value option to volume_to_data
Adding the fill_value argument for the function volume_to_data. It is used to put a placeholder value to vertes that didn't get an attributed voxel value. This can be used with pvalues and the 'vmax' and 'over' agrument in the add_activation method of the BrainOjb to avoid having colored empty area (very useful when displaying pvalues).